# swagger.yaml
openapi: 3.0.0
info:
  title: User API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: http://localhost:3001
    description: Optional server description, e.g. Main (production) server
  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing
paths:
  /user:
    get:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
    post:
      summary: Creates a user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                StudentID:
                  type: number
                Firstname:
                  type: string
                Lastname:
                  type: string
                DOB:
                  type: string
                Mobilephone:
                  type: string
      responses: 
        '201':
          description: Created
    put:
      summary: Update a user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                StudentID:
                  type: number
                Firstname:
                  type: string
                Lastname:
                  type: string
                DOB:
                  type: string
                Mobilephone:
                  type: string
      responses: 
        '201':
          description: Updated
  /user/{StudentID}:
    get:
      summary: Returns a user by ID.
      parameters:
        - name: StudentID
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK
    delete:
      summary: Delete a user by ID.
      parameters:
        - name: StudentID
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK